+++ crates/swc/tests/tsc-references/computedPropertyNames13_ES5_es5.2.minified.js
@@ -1,1 +1,3 @@
+var a;
 import _class_call_check from "@swc/helpers/src/_class_call_check.mjs";
+"hello ".concat(a, " bye");